+++ app/src/main/java/com/wd/woodong2/retrofit/GCMRetrofitClient.kt
@@ -1,33 +1,9 @@
 package com.wd.woodong2.retrofit
 
-import com.wd.woodong2.presentation.home.map.HomeMapSearchRemoteDatasource
 import com.wd.woodong2.services.fcm.GCMService
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-
-object KAKAORetrofitClient {
-
-    private const val BASE_URL = "https://dapi.kakao.com"
-
-    private val okHttpClient by lazy {
-        OkHttpClient.Builder()
-            .addInterceptor(RequestInterceptor())
-            .build()
-    }
-
-    private val retrofit by lazy {
-        Retrofit.Builder()
-            .baseUrl(BASE_URL)
-            .client(okHttpClient)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-    }
-
-    val search: HomeMapSearchRemoteDatasource by lazy {
-        retrofit.create(HomeMapSearchRemoteDatasource::class.java)
-    }
-}
 
 object GCMRetrofitClient {
