+++ app/src/main/java/com/wd/woodong2/retrofit/KAKAORequestInterceptor.kt
@@ -5,7 +5,7 @@ import com.wd.woodong2.WooDongApp
 import okhttp3.Interceptor
 import okhttp3.Response
 
-class RequestInterceptor : Interceptor {
+class KAKAORequestInterceptor : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         val newRequest = chain.request().newBuilder()
             .addHeader(
